+++ R/pmacroModel.R
@@ -5,6 +5,7 @@
 #'@param radx horizontal radius of the box.
 #'@param rady vertical radius of the box.
 #'@param xmargin horizontal margin of plot
+#'@param box.col fill color of box
 #' @param xlim the x limits (min,max) of the plot
 #' @param ylim the y limits (min,max) of the plot
 #' @export
@@ -14,7 +15,7 @@
 #' pmacroModel(1,covar=covar)
 #' covar=list(name=c("C1","C2","C3"),label=c("ese","sex","tenure"),site=list("M",c("Mi","Y"),c("Y")))
 #' pmacroModel(4,covar=covar)
-pmacroModel=function(no=1,labels=list(),covar=list(),radx=0.06,rady=0.06,xmargin=0.03,
+pmacroModel=function(no=1,labels=list(),covar=list(),radx=0.06,rady=0.06,xmargin=0.03,box.col="white",
                      xlim=NULL,ylim=NULL){
     # no=76
     # labels=list("X"="age","M"="educ","Y"="interest","W"="policy","Z"="male")
@@ -83,9 +84,11 @@ pmacroModel=function(no=1,labels=list(),covar=list(),radx=0.06,rady=0.06,xmargin
     }
 
     if(no==74) {
-        conceptDiagram2(X=X,M=M,Y=Y,xb=TRUE,covar=covar,radx=radx,rady=rady,xmargin=xmargin,xlim=xlim,ylim=ylim)
+        conceptDiagram2(X=X,M=M,Y=Y,xb=TRUE,covar=covar,radx=radx,rady=rady,xmargin=xmargin,
+                        box.col=box.col,xlim=xlim,ylim=ylim)
     } else {
-        conceptDiagram2(X=X,M=M,Y=Y,moderator = moderator,covar=covar,radx=radx,rady=rady,xmargin=xmargin,xlim=xlim,ylim=ylim)
+        conceptDiagram2(X=X,M=M,Y=Y,moderator = moderator,covar=covar,radx=radx,rady=rady,xmargin=xmargin,
+                        box.col=box.col,xlim=xlim,ylim=ylim)
     }
 
         }
@@ -94,7 +97,7 @@ pmacroModel=function(no=1,labels=list(),covar=list(),radx=0.06,rady=0.06,xmargin
         straightarrow(c(0.20,0.75),to=c(0.485,0.65),lwd=1,arr.pos=1,arr.type="triangle")
         label=ifelse(is.null(labels$Z),"Z",labels$Z)
 
-        textrect(mid=c(0.20,0.75),radx=radx,rady=rady,lab=label)
+        textrect(mid=c(0.20,0.75),radx=radx,rady=rady,lab=label,box.col=box.col)
     }
     if(no %in% c(11:13,25:27)){
 
@@ -113,7 +116,7 @@ pmacroModel=function(no=1,labels=list(),covar=list(),radx=0.06,rady=0.06,xmargin
         }
         label=ifelse(is.null(labels$Z),"Z",labels$Z)
 
-        textrect(mid=c(radx+xmargin,(0.9+yinterval)/2),radx=radx,rady=rady,lab=label)
+        textrect(mid=c(radx+xmargin,(0.9+yinterval)/2),radx=radx,rady=rady,lab=label,box.col=box.col)
 
     }
     if(no %in% 18:20){
@@ -133,7 +136,7 @@ pmacroModel=function(no=1,labels=list(),covar=list(),radx=0.06,rady=0.06,xmargin
         }
         label=ifelse(is.null(labels$Z),"Z",labels$Z)
 
-        textrect(mid=c(1-(radx+xmargin),(0.9+yinterval)/2),radx=radx,rady=rady,lab=ifelse(is.null(labels$Z),"Z",labels$Z))
+        textrect(mid=c(1-(radx+xmargin),(0.9+yinterval)/2),radx=radx,rady=rady,lab=ifelse(is.null(labels$Z),"Z",labels$Z),box.col=box.col)
 
     }